Traffic-Violation-Detection/requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
easyocr==1.7.1
mysql-connector-python==8.2.0
Pillow==10.1.0
numpy==1.26.4
Werkzeug==3.0.0